Hacer que el campo precio de los tomos sea compatible con los decimales

+++ grails-app/domain/Colecciones/Mangas.groovy
@@ -14,7 +14,7 @@ class Mangas {
 	
 	int numTomosActuales
 	int numTomosMaximos
-	float precioTotal
+	BigDecimal precioTotal
 
 	Date fechaInscripcion
 	Date ultimaModificacion
